Added ideal fulfillment

+++ FindeyVouchers/FindeyVouchers.Domain/PaymentStatusResponse.cs
@@ -4,7 +4,7 @@
     {
         public string PaymentId { get; set; }
         public string PaymentStatus { get; set; }
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public long Created { get; set; }
         public string ErrorMessage { get; set; }
     }
